backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.4
contextlib2==21.6.0
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-ninja==1.5.3
django-ninja-extra==0.31.0
django-ninja-jwt==5.4.4
dnspython==2.8.0
email-validator==2.3.0
idna==3.11
injector==0.24.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pypdfium2==5.5.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
